Add UpdateSettings mutation
- Also some refactorings
