+++ fittrak/users/graphql/user.py
@@ -1,11 +1,12 @@
-""" Users GraphQL schema """
+"""
+Graphql User types
+"""
 
 import graphene
 from django.contrib.auth import get_user_model
 from graphene_django.types import DjangoObjectType
-from graphql import GraphQLError
 
-from .models import Profile
+from ..models import Profile
 
 
 class UnitEnumType(graphene.Enum):
@@ -26,16 +27,3 @@ class Viewer(DjangoObjectType):
     class Meta:
         model = get_user_model()
         only_fields = ("id", "username", "email")
-
-
-class Query:
-    viewer = graphene.Field(Viewer)
-
-    @staticmethod
-    def resolve_viewer(_, info):
-        user = info.context.user
-
-        if not user.is_authenticated:
-            raise GraphQLError("Not authenticated.")
-
-        return user
